+++ CHANGELOG.md
@@ -37,6 +37,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+### API Breaking
+
+- (fees) [\#691](https://github.com/tharsis/evmos/pull/691) Internal Api audit.
+
+## [v5.0.0] - 2022-06-14
+
 ### State Machine Breaking
 
 - (deps) [\#684](https://github.com/tharsis/evmos/pull/684) Bump ibc-go version to [`v3.1.0`](https://github.com/cosmos/ibc-go/releases/tag/v3.1.0)